improve error handling and output
Use a _main helper to easily throw errors with the ? operator.
The main function only has to unwrap the error message, print it and
exit with an error status.
For the moment, only git errors can be returned. In the future, we
probably will need a more complex exception.

Additionally, improve output of the list of commits selected.