add bootstrap command to initialize local repo
The syncing between the two repos is done by matching special tags
in the local repo's commits, so that the matching between commits of the
two repos can be done.

In order to be able to sync, the local repo must first be bootstrapped.
A new option is added for this.