abort if repo has local changes
Check that the repo on which ripit is run does not contain any changes,
either staged or unstaged.

This causes tests to fail because "repo.commit(...)" does not update the
workdir to the commit state. Adding a checkout_head after committing a
file ensures that it does, and allows us to fix a test where a private
file was still present after a bootstrap.