add filtering in commit messages
Add options to declare filters that, if it matches lines in the commit
message, will filter out those lines.

To allow regex expression, a new dependency on the regex crate is added.
This filtering is done by line. If we want in the future to filter out
blocks, this feature should be improved.