sync: use cherrypick method when syncing commits
Call cherrypick to apply the commits changes on the current HEAD.
The previous version imported the whole commit tree, making it possible
to leak previous commits by syncing commits brought by merges.

This requires an update of git2-rs to 0.9.2, where our PR to add the
cherrypick method was merged.