sync: properly sync merge commits
When syncing commits, if we reach a merge commit and we know all its
parents, we want to properly recreate the same merge commit. This is now
done by matching its parents to their local copies, and recreating the
commit with those parents.

We are cherry-picking considering the local branch is the first
parent. I don't think we can really make this hypothesis, so we should
come up with a proper test where this is not true, and handle it
properly when syncing (this is also part of the TODO involving
properly tracking and advancing the local branch when syncing).