sync: allow commits uprooting on unknown parents
When synchronizing a merge commit with one ancestor starting from before
the boostrap, we will try to sync commits with unknown parents, leading
to errors.

To allow this, add a new flag, and "uproot" those commits on HEAD. This
might lead to conflicts due to the break of the topological integrity of
the synchronization, which will be handled in the following commits.