sync: improve addition of tag in commit message
For commit messages with only a brief, the tag was added just after the
brief, without a paragraph change:

brief
rip-it: ...

This cause the tag to be part of the brief, which we do not want.
The addition of the tag is improved, to make sure it is always in its
own paragraph, except if the commit has no message whatsoever.