sync: update branch ref when syncing on top of it
When syncing commits, we either sync a commit on top of the local
branch, or from an ancestor (when rebuilding a merge). When syncing on
top of the local branch, we want to update it. To do that, we detect
if the first parent is the commit tracked by the branch, and update it
if that is the case after the cherrypick.