sync: handle uproot of merge commit
A merge commit cannot be detected by the number of local parents it
has, because it can be uprooted. This can happen if one of the
parent is unknown (was part of the repo prior to the bootstrap), and
a merge commit is brought to the branch being synced by another merge
commit.
In practice, this can happen quite easily if there are several stable
branches, and fixes on those branches are merged progressively into
the current branch:

 (branch v1) fix version v1 --
                              \
 (branch v2) -------------------> merge v1 in v2 -
                                                  \
 (master)    bootstrap -----------------------------> merge v2 in master

Here, "merge v1 in v2" will be uprooted.

To handle this, we need to retrieve the information whether the uprooted
commit is a merge before mapping its parent into the local repo.