sync: update local branch when syncing merges
When syncing a merge, we will first sync the commits of the second
mainline, then synchronize the merge, so we end up with this situation:

           ---> C1 ---
          /           \
remote: C0 --------------> C2

local: C0 [branch]

* sync C1: C0 ---> C1 [branch]

              ---> C1 [branch] ---
             /                    \
* sync C2: C0 ----------------------> C2

The first parent is detached, and the second parent is the local branch.
So the commit must be done detached, then the branch updated.