sync: properly filter commit msg on uprooted commit w/ conflicts
If an uprooted commit has conflicts, we let the user resolve the
conflicts, and do the commit. The commit message is taken from
.git/MERGE_MSG, which is automatically written by git when
cherry-picking the change.

We added the ripit-tag to this file so that the next sync could properly
work, but we forgot to also filter out lines from the commit message.