sync: ignore all already synced commits on remote/local diff
To sync the local repo with the remote one, a revwalk is done between
the last synced commit from the local repo, and last commit of the
remote one.
However, when the last synced commit has been uprooted, the revwalk
should be done from the last non-uprooted synced commit, which is found
by rewinding on the first parent of uprooted commits, until a
non-uprooted one is found. Then, uprooted commits are ignored from the
revwalk, as they have already been synced.

However, it is not enough to ignore the uprooted commits that we found
when rewinding, as others can come from other parents. This can happen
when uprooting a merge commit, but still keeping it as a merge (ie, with
multiple parents).

For example, in this situation:

A1 -------------------->  A2 -                         [branch A]
     \                        \
B1 ----> B2 ----> B3 ------------> B4 -                [branch B]
             \                         \
----------------> M1 (bootstrap) ----------> M2 (sync) [master]

Syncing M2 will uproot A2, B3 and B4. As A2 and B3 will be uprooted and
present in the local repo, B4 will be uprooted but kept as a merge
commit. Syncing from B4 will rewing to the bootstrap, and again list A2,
B3, B4 and M2 as to be synced, but only M2 is to be synced. So A2 has to
be ignored too.

The fix is fairly simple, instead of keeping a set of the uprooted
commits that were rewound, use the commits map (set of all already
synced commits) to find out if a commit should be ignored or not.