commits_map: save map of synced commits in a file
When syncing branches and uprooting commits, we may stumble upon conflicts
and have the user resolve them.
However, when the user runs the ripit tool again, we won't have saved the mapping
of the newly created commit, and it may be in a un-reachable state (not HEAD, not the
local branch).
To solve those type of issues, the state of the mapping between the two
repositories (or rather, the list of synced commits) must be saved
externally.
This is now done by creating a .ripit-cache file in the local
repository, which is reloaded on start-up,
and filled with newly synced commits.