sync: fix creation of uprooted merge commits
When a merge commit is uprooted, its changes are brought with a
cherrypick, so as to avoid bringing its entire tree (and thus leaking
data from before the bootstrap).
However, if there are conflicts, the context saved by git is that of a
cherry-pick, and a "git commit" following conflicts resolution will
create a simple commit, not the merge commit we want.

To fix this, we need to fix the context saved in the .git directory, so
that git understands we are creating a merge commit, with the proper
parents.