Cargo.toml: update version to 0.9
The version is bumped to 0.9 to indicate "beta" status. The tool is
supposed to be mature enough to be able to handle all situations
without having to implement missing features.
Once the last bugs have been fixed (or if stable for long enough),
the version will be bumped to 1.0, and the following updates will
use semantic versioning.