app: add support for multiple branches synchronization
* Add "branches" option in the config file. Support for the "branch"
  option is kept, but is considered legacy.
* Modify functions in sync.rs to work on the "app::Branch" object. There
  are starting to have too many arguments, some rework might be needed.
* in main.rs, iterate on every branch when bootstrapping or syncing.

When bootstrapping, several branches might point to the same commit, in
which case the bootstrap commit should be re-used. This is done by using
the commits map to save & re-use those commits.

There is still an issue when a branch share commits to sync with
another: the sync of the first branch will copy the shared commits,
and the second branch won't be updated to point to them.
This will be fixed in the following commit.